Se agrega rol de usuarios premium

+++ src/app.js
@@ -91,7 +91,6 @@ app.use((req, res, next) => {
     if (req.isAuthenticated()) {
         return next();
     }
-
     return res.redirect('/login');
 });
 
@@ -118,20 +117,30 @@ socketServer.on('connection', socket => {
     //Socket para eliminar productos
     socket.on('eliminar_producto', async (data) => {
         const productId = data.productId;
-
+        const userId = data.userId;
+        const userRole = data.userRole;
+        const productOwner = data.productOwner;
+    
         try {
-
+            if (!userId || !userRole || !productOwner) {
+                console.error('Datos incompletos recibidos para eliminar el producto');
+                return;
+            }
+    
+            if (userRole === 'premium' && productOwner !== userId) {
+                console.error('Usuarios premium solo pueden eliminar sus propios productos');
+                return;
+            }
+    
             await productsRepository.delete(productId);
             io.emit('producto_eliminado', { productId });
-
+    
             const updatedProducts = await productsRepository.getAll();
             io.emit('updateProducts', updatedProducts);
-
         } catch (error) {
             console.error('Error al eliminar el producto:', error);
         }
     });
-
     //Socket para chat:
     socket.on('chatMessage', (data) => {
         const { user, message } = data;
